src/porter_ai_assistant/requirements.txt
# Porter AI Assistant — Python Dependencies
#
# Install in a venv (NOT system Python):
#   python3 -m venv .venv
#   source .venv/bin/activate
#   pip install -r src/porter_ai_assistant/requirements.txt
#
# For RPi 4 ARM64, llama-cpp-python builds from source (needs cmake, gcc).
# Pre-built wheels may be available: pip install llama-cpp-python --prefer-binary

# ── Core Inference ───────────────────────────────────────────────────────────
llama-cpp-python>=0.3.0          # GGUF model loading + inference
PyYAML>=6.0                      # System prompt YAML parsing

# ── Model Download ───────────────────────────────────────────────────────────
huggingface-hub>=0.20.0          # Download models from HuggingFace

# ── Benchmarking & Monitoring ────────────────────────────────────────────────
psutil>=5.9.0                    # Process memory monitoring

# ── Fine-tuning (optional — only needed for training, not inference) ─────────
# Uncomment these for LoRA fine-tuning on a GPU machine:
# torch>=2.1.0
# transformers>=4.38.0
# peft>=0.8.0
# datasets>=2.16.0
# trl>=0.7.0
# bitsandbytes>=0.42.0
# accelerate>=0.25.0
# scipy>=1.11.0
